refactor(core/graph): split incremental module into directory submodules
Split the monolithic graphrag-core/src/graph/incremental.rs (3024 lines)
into a directory module with focused submodules:

- types.rs: core types (ChangeRecord, GraphDelta, IncrementalConfig, etc.)
- monitor.rs: UpdateMonitor and metrics
- change_log.rs: ChangeLog persistence and replay
- conflict.rs: ConflictResolver strategies
- batch.rs: BatchProcessor and PendingBatch
- pagerank.rs: IncrementalPageRank
- mod.rs: IncrementalGraphManager, ProductionGraphStore, traits, public re-exports

Keeps BatchMetrics unconditional so non-incremental builds still compile.

Refs: #28
