--- graphrag-core/src/graph/incremental.rs
@@ -1,3024 +1,0 @@
-//! Incremental graph updates: change tracking, conflict resolution, and selective cache invalidation.
-
-use crate::core::{
-    DocumentId, Entity, EntityId, GraphRAGError, KnowledgeGraph, Relationship, Result, TextChunk,
-};
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
-use std::time::{Duration, Instant};
-
-#[cfg(feature = "incremental")]
-use std::sync::Arc;
-
-#[cfg(feature = "incremental")]
-use {
-    dashmap::DashMap,
-    parking_lot::{Mutex, RwLock},
-    tokio::sync::{broadcast, Semaphore},
-    uuid::Uuid,
-};
-
-// ============================================================================
-// Core Types and Enums
-// ============================================================================
-
-/// Unique identifier for update operations
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct UpdateId(String);
-
-impl UpdateId {
-    /// Creates a new unique update identifier
-    pub fn new() -> Self {
-        #[cfg(feature = "incremental")]
-        {
-            Self(Uuid::new_v4().to_string())
-        }
-        #[cfg(not(feature = "incremental"))]
-        {
-            Self(format!(
-                "update_{}",
-                Utc::now().timestamp_nanos_opt().unwrap_or(0)
-            ))
-        }
-    }
-
-    /// Creates an update identifier from an existing string
-    pub fn from_string(id: String) -> Self {
-        Self(id)
-    }
-
-    /// Returns the update ID as a string slice
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl std::fmt::Display for UpdateId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Default for UpdateId {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Change record for tracking individual modifications
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChangeRecord {
-    /// Unique identifier for this change
-    pub change_id: UpdateId,
-    /// Timestamp when the change occurred
-    pub timestamp: DateTime<Utc>,
-    /// Type of change performed
-    pub change_type: ChangeType,
-    /// Optional entity ID affected by this change
-    pub entity_id: Option<EntityId>,
-    /// Optional document ID affected by this change
-    pub document_id: Option<DocumentId>,
-    /// Operation type (insert, update, delete, upsert)
-    pub operation: Operation,
-    /// Data associated with the change
-    pub data: ChangeData,
-    /// Additional metadata for the change
-    pub metadata: HashMap<String, String>,
-}
-
-/// Types of changes that can occur
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum ChangeType {
-    /// An entity was added to the graph
-    EntityAdded,
-    /// An existing entity was updated
-    EntityUpdated,
-    /// An entity was removed from the graph
-    EntityRemoved,
-    /// A relationship was added to the graph
-    RelationshipAdded,
-    /// An existing relationship was updated
-    RelationshipUpdated,
-    /// A relationship was removed from the graph
-    RelationshipRemoved,
-    /// A document was added
-    DocumentAdded,
-    /// An existing document was updated
-    DocumentUpdated,
-    /// A document was removed
-    DocumentRemoved,
-    /// A text chunk was added
-    ChunkAdded,
-    /// An existing text chunk was updated
-    ChunkUpdated,
-    /// A text chunk was removed
-    ChunkRemoved,
-    /// An embedding was added
-    EmbeddingAdded,
-    /// An existing embedding was updated
-    EmbeddingUpdated,
-    /// An embedding was removed
-    EmbeddingRemoved,
-}
-
-/// Operations that can be performed
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum Operation {
-    /// Insert a new item
-    Insert,
-    /// Update an existing item
-    Update,
-    /// Delete an item
-    Delete,
-    /// Insert or update (upsert) an item
-    Upsert,
-}
-
-/// Data associated with a change
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ChangeData {
-    /// Entity data
-    Entity(Entity),
-    /// Relationship data
-    Relationship(Relationship),
-    /// Document data
-    Document(Document),
-    /// Text chunk data
-    Chunk(Box<TextChunk>),
-    /// Embedding data with entity ID and vector
-    Embedding {
-        /// Entity ID for the embedding
-        entity_id: EntityId,
-        /// Embedding vector
-        embedding: Vec<f32>,
-    },
-    /// Empty change data placeholder
-    Empty,
-}
-
-/// Document type for incremental updates
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Document {
-    /// Unique identifier for the document
-    pub id: DocumentId,
-    /// Document title
-    pub title: String,
-    /// Document content
-    pub content: String,
-    /// Additional metadata
-    pub metadata: HashMap<String, String>,
-}
-
-/// Atomic change set representing a transaction
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GraphDelta {
-    /// Unique identifier for this delta
-    pub delta_id: UpdateId,
-    /// Timestamp when the delta was created
-    pub timestamp: DateTime<Utc>,
-    /// List of changes in this delta
-    pub changes: Vec<ChangeRecord>,
-    /// Delta IDs that this delta depends on
-    pub dependencies: Vec<UpdateId>,
-    /// Current status of the delta
-    pub status: DeltaStatus,
-    /// Data needed to rollback this delta
-    pub rollback_data: Option<RollbackData>,
-}
-
-/// Status of a delta operation
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub enum DeltaStatus {
-    /// Delta is pending application
-    Pending,
-    /// Delta has been applied but not committed
-    Applied,
-    /// Delta has been committed
-    Committed,
-    /// Delta has been rolled back
-    RolledBack,
-    /// Delta failed with error message
-    Failed {
-        /// Error message describing the failure
-        error: String,
-    },
-}
-
-/// Data needed for rollback operations
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RollbackData {
-    /// Previous state of entities before the change
-    pub previous_entities: Vec<Entity>,
-    /// Previous state of relationships before the change
-    pub previous_relationships: Vec<Relationship>,
-    /// Cache keys affected by the change
-    pub affected_caches: Vec<String>,
-}
-
-/// Conflict resolution strategies
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ConflictStrategy {
-    /// Keep the existing data, discard new changes
-    KeepExisting,
-    /// Keep the new data, discard existing
-    KeepNew,
-    /// Merge existing and new data intelligently
-    Merge,
-    /// Use LLM to decide how to resolve conflict
-    LLMDecision,
-    /// Prompt user to resolve conflict
-    UserPrompt,
-    /// Use a custom resolver by name
-    Custom(String),
-}
-
-/// Conflict detected during update
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Conflict {
-    /// Unique identifier for this conflict
-    pub conflict_id: UpdateId,
-    /// Type of conflict detected
-    pub conflict_type: ConflictType,
-    /// Existing data in the graph
-    pub existing_data: ChangeData,
-    /// New data attempting to be applied
-    pub new_data: ChangeData,
-    /// Resolution if already resolved
-    pub resolution: Option<ConflictResolution>,
-}
-
-/// Types of conflicts
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ConflictType {
-    /// Entity already exists with different data
-    EntityExists,
-    /// Relationship already exists with different data
-    RelationshipExists,
-    /// Version mismatch between expected and actual
-    VersionMismatch,
-    /// Data is inconsistent with graph state
-    DataInconsistency,
-    /// Change violates a constraint
-    ConstraintViolation,
-}
-
-/// Resolution for a conflict
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConflictResolution {
-    /// Strategy used to resolve the conflict
-    pub strategy: ConflictStrategy,
-    /// Resolved data after applying strategy
-    pub resolved_data: ChangeData,
-    /// Metadata about the resolution
-    pub metadata: HashMap<String, String>,
-}
-
-// ============================================================================
-// IncrementalGraphStore Trait
-// ============================================================================
-
-/// Extended trait for incremental graph operations with production-ready features
-#[async_trait::async_trait]
-pub trait IncrementalGraphStore: Send + Sync {
-    /// The error type for incremental graph operations
-    type Error: std::error::Error + Send + Sync + 'static;
-
-    /// Upsert an entity (insert or update)
-    async fn upsert_entity(&mut self, entity: Entity) -> Result<UpdateId>;
-
-    /// Upsert a relationship
-    async fn upsert_relationship(&mut self, relationship: Relationship) -> Result<UpdateId>;
-
-    /// Delete an entity and its relationships
-    async fn delete_entity(&mut self, entity_id: &EntityId) -> Result<UpdateId>;
-
-    /// Delete a relationship
-    async fn delete_relationship(
-        &mut self,
-        source: &EntityId,
-        target: &EntityId,
-        relation_type: &str,
-    ) -> Result<UpdateId>;
-
-    /// Apply a batch of changes atomically
-    async fn apply_delta(&mut self, delta: GraphDelta) -> Result<UpdateId>;
-
-    /// Rollback a delta
-    async fn rollback_delta(&mut self, delta_id: &UpdateId) -> Result<()>;
-
-    /// Get change history
-    async fn get_change_log(&self, since: Option<DateTime<Utc>>) -> Result<Vec<ChangeRecord>>;
-
-    /// Start a transaction for atomic operations
-    async fn begin_transaction(&mut self) -> Result<TransactionId>;
-
-    /// Commit a transaction
-    async fn commit_transaction(&mut self, tx_id: TransactionId) -> Result<()>;
-
-    /// Rollback a transaction
-    async fn rollback_transaction(&mut self, tx_id: TransactionId) -> Result<()>;
-
-    /// Batch upsert entities with conflict resolution
-    async fn batch_upsert_entities(
-        &mut self,
-        entities: Vec<Entity>,
-        _strategy: ConflictStrategy,
-    ) -> Result<Vec<UpdateId>>;
-
-    /// Batch upsert relationships with conflict resolution
-    async fn batch_upsert_relationships(
-        &mut self,
-        relationships: Vec<Relationship>,
-        _strategy: ConflictStrategy,
-    ) -> Result<Vec<UpdateId>>;
-
-    /// Update entity embeddings incrementally
-    async fn update_entity_embedding(
-        &mut self,
-        entity_id: &EntityId,
-        embedding: Vec<f32>,
-    ) -> Result<UpdateId>;
-
-    /// Bulk update embeddings for performance
-    async fn bulk_update_embeddings(
-        &mut self,
-        updates: Vec<(EntityId, Vec<f32>)>,
-    ) -> Result<Vec<UpdateId>>;
-
-    /// Get pending transactions
-    async fn get_pending_transactions(&self) -> Result<Vec<TransactionId>>;
-
-    /// Get graph statistics
-    async fn get_graph_statistics(&self) -> Result<GraphStatistics>;
-
-    /// Validate graph consistency
-    async fn validate_consistency(&self) -> Result<ConsistencyReport>;
-}
-
-/// Transaction identifier for atomic operations
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct TransactionId(String);
-
-impl TransactionId {
-    /// Creates a new unique transaction identifier
-    pub fn new() -> Self {
-        #[cfg(feature = "incremental")]
-        {
-            Self(Uuid::new_v4().to_string())
-        }
-        #[cfg(not(feature = "incremental"))]
-        {
-            Self(format!(
-                "tx_{}",
-                Utc::now().timestamp_nanos_opt().unwrap_or(0)
-            ))
-        }
-    }
-
-    /// Returns the transaction ID as a string slice
-    pub fn as_str(&self) -> &str {
-        &self.0
-    }
-}
-
-impl std::fmt::Display for TransactionId {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl Default for TransactionId {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Graph statistics for monitoring
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GraphStatistics {
-    /// Total number of nodes (entities)
-    pub node_count: usize,
-    /// Total number of edges (relationships)
-    pub edge_count: usize,
-    /// Average degree of nodes
-    pub average_degree: f64,
-    /// Maximum degree of any node
-    pub max_degree: usize,
-    /// Number of connected components
-    pub connected_components: usize,
-    /// Clustering coefficient
-    pub clustering_coefficient: f64,
-    /// When statistics were last updated
-    pub last_updated: DateTime<Utc>,
-}
-
-/// Consistency validation report
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ConsistencyReport {
-    /// Whether the graph is consistent
-    pub is_consistent: bool,
-    /// Entities with no relationships
-    pub orphaned_entities: Vec<EntityId>,
-    /// Relationships referencing non-existent entities
-    pub broken_relationships: Vec<(EntityId, EntityId, String)>,
-    /// Entities missing embeddings
-    pub missing_embeddings: Vec<EntityId>,
-    /// When validation was performed
-    pub validation_time: DateTime<Utc>,
-    /// Total number of issues found
-    pub issues_found: usize,
-}
-
-// ============================================================================
-// Cache Management
-// ============================================================================
-
-/// Cache invalidation strategies
-#[derive(Debug, Clone)]
-pub enum InvalidationStrategy {
-    /// Invalidate specific cache keys
-    Selective(Vec<String>),
-    /// Invalidate all caches in a region
-    Regional(String),
-    /// Invalidate all caches
-    Global,
-    /// Invalidate based on entity relationships
-    Relational(EntityId, u32), // entity_id, depth
-}
-
-/// Cache region affected by changes
-#[derive(Debug, Clone)]
-pub struct CacheRegion {
-    /// Unique identifier for the cache region
-    pub region_id: String,
-    /// Entity IDs in this region
-    pub entity_ids: HashSet<EntityId>,
-    /// Relationship types in this region
-    pub relationship_types: HashSet<String>,
-    /// Document IDs in this region
-    pub document_ids: HashSet<DocumentId>,
-    /// When the region was last modified
-    pub last_modified: DateTime<Utc>,
-}
-
-/// Selective cache invalidation manager
-#[cfg(feature = "incremental")]
-pub struct SelectiveInvalidation {
-    cache_regions: DashMap<String, CacheRegion>,
-    entity_to_regions: DashMap<EntityId, HashSet<String>>,
-    invalidation_log: Mutex<Vec<(DateTime<Utc>, InvalidationStrategy)>>,
-}
-
-#[cfg(feature = "incremental")]
-impl Default for SelectiveInvalidation {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(feature = "incremental")]
-impl SelectiveInvalidation {
-    /// Creates a new selective invalidation manager
-    pub fn new() -> Self {
-        Self {
-            cache_regions: DashMap::new(),
-            entity_to_regions: DashMap::new(),
-            invalidation_log: Mutex::new(Vec::new()),
-        }
-    }
-
-    /// Registers a cache region for invalidation tracking
-    pub fn register_cache_region(&self, region: CacheRegion) {
-        let region_id = region.region_id.clone();
-
-        // Update entity mappings
-        for entity_id in &region.entity_ids {
-            self.entity_to_regions
-                .entry(entity_id.clone())
-                .or_default()
-                .insert(region_id.clone());
-        }
-
-        self.cache_regions.insert(region_id, region);
-    }
-
-    /// Determines invalidation strategies for a set of changes
-    pub fn invalidate_for_changes(&self, changes: &[ChangeRecord]) -> Vec<InvalidationStrategy> {
-        let mut strategies = Vec::new();
-        let mut affected_regions = HashSet::new();
-
-        for change in changes {
-            match &change.change_type {
-                ChangeType::EntityAdded | ChangeType::EntityUpdated | ChangeType::EntityRemoved => {
-                    if let Some(entity_id) = &change.entity_id {
-                        if let Some(regions) = self.entity_to_regions.get(entity_id) {
-                            affected_regions.extend(regions.clone());
-                        }
-                        strategies.push(InvalidationStrategy::Relational(entity_id.clone(), 2));
-                    }
-                },
-                ChangeType::RelationshipAdded
-                | ChangeType::RelationshipUpdated
-                | ChangeType::RelationshipRemoved => {
-                    // Invalidate based on relationship endpoints
-                    if let ChangeData::Relationship(rel) = &change.data {
-                        strategies.push(InvalidationStrategy::Relational(rel.source.clone(), 1));
-                        strategies.push(InvalidationStrategy::Relational(rel.target.clone(), 1));
-                    }
-                },
-                _ => {
-                    // For other changes, use selective invalidation
-                    let cache_keys = self.generate_cache_keys_for_change(change);
-                    if !cache_keys.is_empty() {
-                        strategies.push(InvalidationStrategy::Selective(cache_keys));
-                    }
-                },
-            }
-        }
-
-        // Add regional invalidation for affected regions
-        for region_id in affected_regions {
-            strategies.push(InvalidationStrategy::Regional(region_id));
-        }
-
-        // Log invalidation
-        let mut log = self.invalidation_log.lock();
-        for strategy in &strategies {
-            log.push((Utc::now(), strategy.clone()));
-        }
-
-        strategies
-    }
-
-    fn generate_cache_keys_for_change(&self, change: &ChangeRecord) -> Vec<String> {
-        let mut keys = Vec::new();
-
-        // Generate cache keys based on change type and data
-        match &change.change_type {
-            ChangeType::EntityAdded | ChangeType::EntityUpdated => {
-                if let Some(entity_id) = &change.entity_id {
-                    keys.push(format!("entity:{entity_id}"));
-                    keys.push(format!("entity_neighbors:{entity_id}"));
-                }
-            },
-            ChangeType::DocumentAdded | ChangeType::DocumentUpdated => {
-                if let Some(doc_id) = &change.document_id {
-                    keys.push(format!("document:{doc_id}"));
-                    keys.push(format!("document_chunks:{doc_id}"));
-                }
-            },
-            ChangeType::EmbeddingAdded | ChangeType::EmbeddingUpdated => {
-                if let Some(entity_id) = &change.entity_id {
-                    keys.push(format!("embedding:{entity_id}"));
-                    keys.push(format!("similarity:{entity_id}"));
-                }
-            },
-            _ => {},
-        }
-
-        keys
-    }
-
-    /// Gets statistics about cache invalidations
-    pub fn get_invalidation_stats(&self) -> InvalidationStats {
-        let log = self.invalidation_log.lock();
-
-        InvalidationStats {
-            total_invalidations: log.len(),
-            cache_regions: self.cache_regions.len(),
-            entity_mappings: self.entity_to_regions.len(),
-            last_invalidation: log.last().map(|(time, _)| *time),
-        }
-    }
-}
-
-/// Statistics about cache invalidations
-#[derive(Debug, Clone)]
-pub struct InvalidationStats {
-    /// Total number of invalidations performed
-    pub total_invalidations: usize,
-    /// Number of cache regions registered
-    pub cache_regions: usize,
-    /// Number of entity-to-region mappings
-    pub entity_mappings: usize,
-    /// Timestamp of last invalidation
-    pub last_invalidation: Option<DateTime<Utc>>,
-}
-
-// ============================================================================
-// Conflict Resolution
-// ============================================================================
-
-/// Conflict resolver with multiple strategies
-pub struct ConflictResolver {
-    strategy: ConflictStrategy,
-    custom_resolvers: HashMap<String, ConflictResolverFn>,
-}
-
-// Reduce type complexity for custom resolver function type
-type ConflictResolverFn = Box<dyn Fn(&Conflict) -> Result<ConflictResolution> + Send + Sync>;
-
-impl ConflictResolver {
-    /// Creates a new conflict resolver with the given strategy
-    pub fn new(strategy: ConflictStrategy) -> Self {
-        Self {
-            strategy,
-            custom_resolvers: HashMap::new(),
-        }
-    }
-
-    /// Adds a custom resolver function by name
-    pub fn with_custom_resolver<F>(mut self, name: String, resolver: F) -> Self
-    where
-        F: Fn(&Conflict) -> Result<ConflictResolution> + Send + Sync + 'static,
-    {
-        self.custom_resolvers.insert(name, Box::new(resolver));
-        self
-    }
-
-    /// Resolves a conflict using the configured strategy
-    pub async fn resolve_conflict(&self, conflict: &Conflict) -> Result<ConflictResolution> {
-        match &self.strategy {
-            ConflictStrategy::KeepExisting => Ok(ConflictResolution {
-                strategy: ConflictStrategy::KeepExisting,
-                resolved_data: conflict.existing_data.clone(),
-                metadata: HashMap::new(),
-            }),
-            ConflictStrategy::KeepNew => Ok(ConflictResolution {
-                strategy: ConflictStrategy::KeepNew,
-                resolved_data: conflict.new_data.clone(),
-                metadata: HashMap::new(),
-            }),
-            ConflictStrategy::Merge => self.merge_conflict_data(conflict).await,
-            ConflictStrategy::Custom(resolver_name) => {
-                if let Some(resolver) = self.custom_resolvers.get(resolver_name) {
-                    resolver(conflict)
-                } else {
-                    Err(GraphRAGError::ConflictResolution {
-                        message: format!("Custom resolver '{resolver_name}' not found"),
-                    })
-                }
-            },
-            _ => Err(GraphRAGError::ConflictResolution {
-                message: "Conflict resolution strategy not implemented".to_string(),
-            }),
-        }
-    }
-
-    async fn merge_conflict_data(&self, conflict: &Conflict) -> Result<ConflictResolution> {
-        match (&conflict.existing_data, &conflict.new_data) {
-            (ChangeData::Entity(existing), ChangeData::Entity(new)) => {
-                let merged = self.merge_entities(existing, new)?;
-                Ok(ConflictResolution {
-                    strategy: ConflictStrategy::Merge,
-                    resolved_data: ChangeData::Entity(merged),
-                    metadata: [("merge_strategy".to_string(), "entity_merge".to_string())]
-                        .into_iter()
-                        .collect(),
-                })
-            },
-            (ChangeData::Relationship(existing), ChangeData::Relationship(new)) => {
-                let merged = self.merge_relationships(existing, new)?;
-                Ok(ConflictResolution {
-                    strategy: ConflictStrategy::Merge,
-                    resolved_data: ChangeData::Relationship(merged),
-                    metadata: [(
-                        "merge_strategy".to_string(),
-                        "relationship_merge".to_string(),
-                    )]
-                    .into_iter()
-                    .collect(),
-                })
-            },
-            _ => Err(GraphRAGError::ConflictResolution {
-                message: "Cannot merge incompatible data types".to_string(),
-            }),
-        }
-    }
-
-    fn merge_entities(&self, existing: &Entity, new: &Entity) -> Result<Entity> {
-        let mut merged = existing.clone();
-
-        // Use higher confidence
-        if new.confidence > existing.confidence {
-            merged.confidence = new.confidence;
-            merged.name = new.name.clone();
-            merged.entity_type = new.entity_type.clone();
-        }
-
-        // Merge mentions
-        let mut all_mentions = existing.mentions.clone();
-        for new_mention in &new.mentions {
-            if !all_mentions.iter().any(|m| {
-                m.chunk_id == new_mention.chunk_id && m.start_offset == new_mention.start_offset
-            }) {
-                all_mentions.push(new_mention.clone());
-            }
-        }
-        merged.mentions = all_mentions;
-
-        // Prefer new embedding if available
-        if new.embedding.is_some() {
-            merged.embedding = new.embedding.clone();
-        }
-
-        Ok(merged)
-    }
-
-    fn merge_relationships(
-        &self,
-        existing: &Relationship,
-        new: &Relationship,
-    ) -> Result<Relationship> {
-        let mut merged = existing.clone();
-
-        // Use higher confidence
-        if new.confidence > existing.confidence {
-            merged.confidence = new.confidence;
-            merged.relation_type = new.relation_type.clone();
-        }
-
-        // Merge contexts
-        let mut all_contexts = existing.context.clone();
-        for new_context in &new.context {
-            if !all_contexts.contains(new_context) {
-                all_contexts.push(new_context.clone());
-            }
-        }
-        merged.context = all_contexts;
-
-        Ok(merged)
-    }
-}
-
-// ============================================================================
-// Update Monitor and Metrics
-// ============================================================================
-
-/// Monitor for tracking update operations and performance
-#[cfg(feature = "incremental")]
-pub struct UpdateMonitor {
-    metrics: DashMap<String, UpdateMetric>,
-    operations_log: Mutex<Vec<OperationLog>>,
-    performance_stats: RwLock<PerformanceStats>,
-}
-
-#[cfg(feature = "incremental")]
-impl Default for UpdateMonitor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Metric for tracking update operations
-#[derive(Debug, Clone)]
-pub struct UpdateMetric {
-    /// Name of the metric
-    pub name: String,
-    /// Metric value
-    pub value: f64,
-    /// When the metric was recorded
-    pub timestamp: DateTime<Utc>,
-    /// Tags for categorizing the metric
-    pub tags: HashMap<String, String>,
-}
-
-/// Log entry for an operation
-#[derive(Debug, Clone)]
-pub struct OperationLog {
-    /// Unique operation identifier
-    pub operation_id: UpdateId,
-    /// Type of operation performed
-    pub operation_type: String,
-    /// When the operation started
-    pub start_time: Instant,
-    /// When the operation ended
-    pub end_time: Option<Instant>,
-    /// Whether the operation succeeded
-    pub success: Option<bool>,
-    /// Error message if failed
-    pub error_message: Option<String>,
-    /// Number of entities affected
-    pub affected_entities: usize,
-    /// Number of relationships affected
-    pub affected_relationships: usize,
-}
-
-/// Performance statistics for monitoring
-#[derive(Debug, Clone)]
-pub struct PerformanceStats {
-    /// Total number of operations performed
-    pub total_operations: u64,
-    /// Number of successful operations
-    pub successful_operations: u64,
-    /// Number of failed operations
-    pub failed_operations: u64,
-    /// Average time per operation
-    pub average_operation_time: Duration,
-    /// Peak throughput in operations per second
-    pub peak_operations_per_second: f64,
-    /// Cache hit rate (0.0 to 1.0)
-    pub cache_hit_rate: f64,
-    /// Conflict resolution rate (0.0 to 1.0)
-    pub conflict_resolution_rate: f64,
-}
-
-#[cfg(feature = "incremental")]
-impl UpdateMonitor {
-    /// Creates a new update monitor
-    pub fn new() -> Self {
-        Self {
-            metrics: DashMap::new(),
-            operations_log: Mutex::new(Vec::new()),
-            performance_stats: RwLock::new(PerformanceStats {
-                total_operations: 0,
-                successful_operations: 0,
-                failed_operations: 0,
-                average_operation_time: Duration::from_millis(0),
-                peak_operations_per_second: 0.0,
-                cache_hit_rate: 0.0,
-                conflict_resolution_rate: 0.0,
-            }),
-        }
-    }
-
-    /// Starts tracking a new operation and returns its ID
-    pub fn start_operation(&self, operation_type: &str) -> UpdateId {
-        let operation_id = UpdateId::new();
-        let log_entry = OperationLog {
-            operation_id: operation_id.clone(),
-            operation_type: operation_type.to_string(),
-            start_time: Instant::now(),
-            end_time: None,
-            success: None,
-            error_message: None,
-            affected_entities: 0,
-            affected_relationships: 0,
-        };
-
-        self.operations_log.lock().push(log_entry);
-        operation_id
-    }
-
-    /// Marks an operation as complete with results
-    pub fn complete_operation(
-        &self,
-        operation_id: &UpdateId,
-        success: bool,
-        error: Option<String>,
-        affected_entities: usize,
-        affected_relationships: usize,
-    ) {
-        // parking_lot::Mutex is not reentrant; release the operations_log
-        // guard before update_performance_stats reacquires it.
-        {
-            let mut log = self.operations_log.lock();
-            if let Some(entry) = log.iter_mut().find(|e| &e.operation_id == operation_id) {
-                entry.end_time = Some(Instant::now());
-                entry.success = Some(success);
-                entry.error_message = error;
-                entry.affected_entities = affected_entities;
-                entry.affected_relationships = affected_relationships;
-            }
-        }
-
-        self.update_performance_stats();
-    }
-
-    fn update_performance_stats(&self) {
-        let log = self.operations_log.lock();
-        let completed_ops: Vec<_> = log
-            .iter()
-            .filter(|op| op.end_time.is_some() && op.success.is_some())
-            .collect();
-
-        if completed_ops.is_empty() {
-            return;
-        }
-
-        let mut stats = self.performance_stats.write();
-        stats.total_operations = completed_ops.len() as u64;
-        stats.successful_operations = completed_ops
-            .iter()
-            .filter(|op| op.success == Some(true))
-            .count() as u64;
-        stats.failed_operations = stats.total_operations - stats.successful_operations;
-
-        // Calculate average operation time
-        let total_time: Duration = completed_ops
-            .iter()
-            .filter_map(|op| op.end_time.map(|end| end.duration_since(op.start_time)))
-            .sum();
-
-        if !completed_ops.is_empty() {
-            stats.average_operation_time = total_time / completed_ops.len() as u32;
-        }
-    }
-
-    /// Records a metric with tags
-    pub fn record_metric(&self, name: &str, value: f64, tags: HashMap<String, String>) {
-        let metric = UpdateMetric {
-            name: name.to_string(),
-            value,
-            timestamp: Utc::now(),
-            tags,
-        };
-        self.metrics.insert(name.to_string(), metric);
-    }
-
-    /// Gets the current performance statistics
-    pub fn get_performance_stats(&self) -> PerformanceStats {
-        self.performance_stats.read().clone()
-    }
-
-    /// Gets the most recent operations up to the specified limit
-    pub fn get_recent_operations(&self, limit: usize) -> Vec<OperationLog> {
-        let log = self.operations_log.lock();
-        log.iter().rev().take(limit).cloned().collect()
-    }
-}
-
-// ============================================================================
-// Main Incremental Graph Manager
-// ============================================================================
-
-/// Comprehensive incremental graph manager with production features
-#[cfg(feature = "incremental")]
-pub struct IncrementalGraphManager {
-    graph: Arc<RwLock<KnowledgeGraph>>,
-    change_log: DashMap<UpdateId, ChangeRecord>,
-    deltas: DashMap<UpdateId, GraphDelta>,
-    cache_invalidation: Arc<SelectiveInvalidation>,
-    conflict_resolver: Arc<ConflictResolver>,
-    monitor: Arc<UpdateMonitor>,
-    config: IncrementalConfig,
-}
-
-#[cfg(not(feature = "incremental"))]
-/// Incremental graph manager (simplified version without incremental feature)
-pub struct IncrementalGraphManager {
-    graph: KnowledgeGraph,
-    change_log: Vec<ChangeRecord>,
-    config: IncrementalConfig,
-}
-
-/// Configuration for incremental operations
-#[derive(Debug, Clone)]
-pub struct IncrementalConfig {
-    /// Maximum number of changes to keep in the log
-    pub max_change_log_size: usize,
-    /// Maximum number of changes in a single delta
-    pub max_delta_size: usize,
-    /// Default conflict resolution strategy
-    pub conflict_strategy: ConflictStrategy,
-    /// Whether to enable performance monitoring
-    pub enable_monitoring: bool,
-    /// Cache invalidation strategy name
-    pub cache_invalidation_strategy: String,
-    /// Default batch size for batch operations
-    pub batch_size: usize,
-    /// Maximum number of concurrent operations
-    pub max_concurrent_operations: usize,
-}
-
-impl Default for IncrementalConfig {
-    fn default() -> Self {
-        Self {
-            max_change_log_size: 10000,
-            max_delta_size: 1000,
-            conflict_strategy: ConflictStrategy::Merge,
-            enable_monitoring: true,
-            cache_invalidation_strategy: "selective".to_string(),
-            batch_size: 100,
-            max_concurrent_operations: 10,
-        }
-    }
-}
-
-#[cfg(feature = "incremental")]
-impl IncrementalGraphManager {
-    /// Creates a new incremental graph manager with feature-gated capabilities
-    pub fn new(graph: KnowledgeGraph, config: IncrementalConfig) -> Self {
-        Self {
-            graph: Arc::new(RwLock::new(graph)),
-            change_log: DashMap::new(),
-            deltas: DashMap::new(),
-            cache_invalidation: Arc::new(SelectiveInvalidation::new()),
-            conflict_resolver: Arc::new(ConflictResolver::new(config.conflict_strategy.clone())),
-            monitor: Arc::new(UpdateMonitor::new()),
-            config,
-        }
-    }
-
-    /// Sets a custom conflict resolver for the manager
-    pub fn with_conflict_resolver(mut self, resolver: ConflictResolver) -> Self {
-        self.conflict_resolver = Arc::new(resolver);
-        self
-    }
-
-    /// Get a read-only reference to the knowledge graph
-    pub fn graph(&self) -> Arc<RwLock<KnowledgeGraph>> {
-        Arc::clone(&self.graph)
-    }
-
-    /// Get the conflict resolver
-    pub fn conflict_resolver(&self) -> Arc<ConflictResolver> {
-        Arc::clone(&self.conflict_resolver)
-    }
-
-    /// Get the update monitor
-    pub fn monitor(&self) -> Arc<UpdateMonitor> {
-        Arc::clone(&self.monitor)
-    }
-}
-
-#[cfg(not(feature = "incremental"))]
-impl IncrementalGraphManager {
-    /// Creates a new incremental graph manager without advanced features
-    pub fn new(graph: KnowledgeGraph, config: IncrementalConfig) -> Self {
-        Self {
-            graph,
-            change_log: Vec::new(),
-            config,
-        }
-    }
-
-    /// Gets a reference to the knowledge graph
-    pub fn graph(&self) -> &KnowledgeGraph {
-        &self.graph
-    }
-
-    /// Gets a mutable reference to the knowledge graph
-    pub fn graph_mut(&mut self) -> &mut KnowledgeGraph {
-        &mut self.graph
-    }
-}
-
-// Common implementation for both feature-gated and non-feature-gated versions
-impl IncrementalGraphManager {
-    /// Create a new change record
-    pub fn create_change_record(
-        &self,
-        change_type: ChangeType,
-        operation: Operation,
-        change_data: ChangeData,
-        entity_id: Option<EntityId>,
-        document_id: Option<DocumentId>,
-    ) -> ChangeRecord {
-        ChangeRecord {
-            change_id: UpdateId::new(),
-            timestamp: Utc::now(),
-            change_type,
-            entity_id,
-            document_id,
-            operation,
-            data: change_data,
-            metadata: HashMap::new(),
-        }
-    }
-
-    /// Get configuration
-    pub fn config(&self) -> &IncrementalConfig {
-        &self.config
-    }
-
-    /// Basic entity upsert (works without incremental feature)
-    pub fn basic_upsert_entity(&mut self, entity: Entity) -> Result<UpdateId> {
-        let update_id = UpdateId::new();
-
-        #[cfg(feature = "incremental")]
-        {
-            let operation_id = self.monitor.start_operation("upsert_entity");
-            let mut graph = self.graph.write();
-
-            match graph.add_entity(entity.clone()) {
-                Ok(_) => {
-                    let ent_id = entity.id.clone();
-                    let change = self.create_change_record(
-                        ChangeType::EntityAdded,
-                        Operation::Upsert,
-                        ChangeData::Entity(entity),
-                        Some(ent_id),
-                        None,
-                    );
-                    self.change_log.insert(change.change_id.clone(), change);
-                    self.monitor
-                        .complete_operation(&operation_id, true, None, 1, 0);
-                    Ok(update_id)
-                },
-                Err(e) => {
-                    self.monitor.complete_operation(
-                        &operation_id,
-                        false,
-                        Some(e.to_string()),
-                        0,
-                        0,
-                    );
-                    Err(e)
-                },
-            }
-        }
-
-        #[cfg(not(feature = "incremental"))]
-        {
-            self.graph.add_entity(entity.clone())?;
-            // Capture ID before moving `entity` into ChangeData
-            let ent_id = entity.id.clone();
-            let change = self.create_change_record(
-                ChangeType::EntityAdded,
-                Operation::Upsert,
-                ChangeData::Entity(entity),
-                Some(ent_id),
-                None,
-            );
-            self.change_log.push(change);
-            Ok(update_id)
-        }
-    }
-}
-
-// ============================================================================
-// Statistics and Monitoring
-// ============================================================================
-
-/// Comprehensive statistics for incremental operations
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct IncrementalStatistics {
-    /// Total number of update operations
-    pub total_updates: usize,
-    /// Number of successful updates
-    pub successful_updates: usize,
-    /// Number of failed updates
-    pub failed_updates: usize,
-    /// Number of entities added
-    pub entities_added: usize,
-    /// Number of entities updated
-    pub entities_updated: usize,
-    /// Number of entities removed
-    pub entities_removed: usize,
-    /// Number of relationships added
-    pub relationships_added: usize,
-    /// Number of relationships updated
-    pub relationships_updated: usize,
-    /// Number of relationships removed
-    pub relationships_removed: usize,
-    /// Number of conflicts resolved
-    pub conflicts_resolved: usize,
-    /// Number of cache invalidations performed
-    pub cache_invalidations: usize,
-    /// Average update time in milliseconds
-    pub average_update_time_ms: f64,
-    /// Peak updates per second achieved
-    pub peak_updates_per_second: f64,
-    /// Current size of the change log
-    pub current_change_log_size: usize,
-    /// Current number of active deltas
-    pub current_delta_count: usize,
-}
-
-impl IncrementalStatistics {
-    /// Creates an empty statistics instance
-    pub fn empty() -> Self {
-        Self {
-            total_updates: 0,
-            successful_updates: 0,
-            failed_updates: 0,
-            entities_added: 0,
-            entities_updated: 0,
-            entities_removed: 0,
-            relationships_added: 0,
-            relationships_updated: 0,
-            relationships_removed: 0,
-            conflicts_resolved: 0,
-            cache_invalidations: 0,
-            average_update_time_ms: 0.0,
-            peak_updates_per_second: 0.0,
-            current_change_log_size: 0,
-            current_delta_count: 0,
-        }
-    }
-
-    /// Prints statistics to stdout in a formatted way
-    pub fn print(&self) {
-        println!("🔄 Incremental Updates Statistics");
-        println!("  Total updates: {}", self.total_updates);
-        println!(
-            "  Successful: {} ({:.1}%)",
-            self.successful_updates,
-            if self.total_updates > 0 {
-                (self.successful_updates as f64 / self.total_updates as f64) * 100.0
-            } else {
-                0.0
-            }
-        );
-        println!("  Failed: {}", self.failed_updates);
-        println!(
-            "  Entities: +{} ~{} -{}",
-            self.entities_added, self.entities_updated, self.entities_removed
-        );
-        println!(
-            "  Relationships: +{} ~{} -{}",
-            self.relationships_added, self.relationships_updated, self.relationships_removed
-        );
-        println!("  Conflicts resolved: {}", self.conflicts_resolved);
-        println!("  Cache invalidations: {}", self.cache_invalidations);
-        println!("  Avg update time: {:.2}ms", self.average_update_time_ms);
-        println!("  Peak updates/sec: {:.1}", self.peak_updates_per_second);
-        println!("  Change log size: {}", self.current_change_log_size);
-        println!("  Active deltas: {}", self.current_delta_count);
-    }
-}
-
-#[cfg(feature = "incremental")]
-impl IncrementalGraphManager {
-    /// Gets comprehensive statistics about incremental operations
-    pub fn get_statistics(&self) -> IncrementalStatistics {
-        let perf_stats = self.monitor.get_performance_stats();
-        let invalidation_stats = self.cache_invalidation.get_invalidation_stats();
-
-        // Calculate entity/relationship statistics from change log
-        let mut entity_stats = (0, 0, 0); // added, updated, removed
-        let mut relationship_stats = (0, 0, 0);
-        let conflicts_resolved = 0;
-
-        for change in self.change_log.iter() {
-            match change.value().change_type {
-                ChangeType::EntityAdded => entity_stats.0 += 1,
-                ChangeType::EntityUpdated => entity_stats.1 += 1,
-                ChangeType::EntityRemoved => entity_stats.2 += 1,
-                ChangeType::RelationshipAdded => relationship_stats.0 += 1,
-                ChangeType::RelationshipUpdated => relationship_stats.1 += 1,
-                ChangeType::RelationshipRemoved => relationship_stats.2 += 1,
-                _ => {},
-            }
-        }
-
-        IncrementalStatistics {
-            total_updates: perf_stats.total_operations as usize,
-            successful_updates: perf_stats.successful_operations as usize,
-            failed_updates: perf_stats.failed_operations as usize,
-            entities_added: entity_stats.0,
-            entities_updated: entity_stats.1,
-            entities_removed: entity_stats.2,
-            relationships_added: relationship_stats.0,
-            relationships_updated: relationship_stats.1,
-            relationships_removed: relationship_stats.2,
-            conflicts_resolved,
-            cache_invalidations: invalidation_stats.total_invalidations,
-            average_update_time_ms: perf_stats.average_operation_time.as_millis() as f64,
-            peak_updates_per_second: perf_stats.peak_operations_per_second,
-            current_change_log_size: self.change_log.len(),
-            current_delta_count: self.deltas.len(),
-        }
-    }
-}
-
-#[cfg(not(feature = "incremental"))]
-impl IncrementalGraphManager {
-    /// Gets basic statistics about incremental operations (non-feature version)
-    pub fn get_statistics(&self) -> IncrementalStatistics {
-        let mut stats = IncrementalStatistics::empty();
-        stats.current_change_log_size = self.change_log.len();
-
-        for change in &self.change_log {
-            match change.change_type {
-                ChangeType::EntityAdded => stats.entities_added += 1,
-                ChangeType::EntityUpdated => stats.entities_updated += 1,
-                ChangeType::EntityRemoved => stats.entities_removed += 1,
-                ChangeType::RelationshipAdded => stats.relationships_added += 1,
-                ChangeType::RelationshipUpdated => stats.relationships_updated += 1,
-                ChangeType::RelationshipRemoved => stats.relationships_removed += 1,
-                _ => {},
-            }
-        }
-
-        stats.total_updates = self.change_log.len();
-        stats.successful_updates = self.change_log.len(); // Assume all succeeded in basic mode
-        stats
-    }
-}
-
-// ============================================================================
-// Incremental PageRank Implementation
-// ============================================================================
-
-/// Incremental PageRank calculator for efficient updates
-#[cfg(feature = "incremental")]
-pub struct IncrementalPageRank {
-    scores: DashMap<EntityId, f64>,
-    // Reserved for delta-PageRank propagation; not yet consumed by `update_incremental`.
-    #[allow(dead_code)]
-    adjacency_changes: DashMap<EntityId, Vec<(EntityId, f64)>>,
-    damping_factor: f64,
-    tolerance: f64,
-    max_iterations: usize,
-    // Tracked for periodic full-recompute heuristics; reader is pending.
-    #[allow(dead_code)]
-    last_full_computation: DateTime<Utc>,
-    incremental_threshold: usize,
-    pending_changes: RwLock<usize>,
-}
-
-#[cfg(feature = "incremental")]
-impl IncrementalPageRank {
-    /// Creates a new incremental PageRank calculator
-    pub fn new(damping_factor: f64, tolerance: f64, max_iterations: usize) -> Self {
-        Self {
-            scores: DashMap::new(),
-            adjacency_changes: DashMap::new(),
-            damping_factor,
-            tolerance,
-            max_iterations,
-            last_full_computation: Utc::now(),
-            incremental_threshold: 1000,
-            pending_changes: RwLock::new(0),
-        }
-    }
-
-    /// Update PageRank incrementally for a specific subgraph
-    pub async fn update_incremental(
-        &self,
-        changed_entities: &[EntityId],
-        graph: &KnowledgeGraph,
-    ) -> Result<()> {
-        let start = Instant::now();
-
-        // If too many changes accumulated, do full recomputation
-        {
-            let pending = *self.pending_changes.read();
-            if pending > self.incremental_threshold {
-                return self.full_recomputation(graph).await;
-            }
-        }
-
-        // Incremental update for changed entities and their neighborhoods
-        let mut affected_entities = HashSet::new();
-
-        // Add changed entities and their neighbors (2-hop neighborhood)
-        for entity_id in changed_entities {
-            affected_entities.insert(entity_id.clone());
-
-            // Add direct neighbors
-            for (neighbor, _) in graph.get_neighbors(entity_id) {
-                affected_entities.insert(neighbor.id.clone());
-
-                // Add second-hop neighbors
-                for (second_hop, _) in graph.get_neighbors(&neighbor.id) {
-                    affected_entities.insert(second_hop.id.clone());
-                }
-            }
-        }
-
-        // Perform localized PageRank computation
-        self.localized_pagerank(&affected_entities, graph).await?;
-
-        // Reset pending changes counter
-        *self.pending_changes.write() = 0;
-
-        let duration = start.elapsed();
-        println!(
-            "🔄 Incremental PageRank update completed in {:?} for {} entities",
-            duration,
-            affected_entities.len()
-        );
-
-        Ok(())
-    }
-
-    /// Perform full PageRank recomputation
-    async fn full_recomputation(&self, graph: &KnowledgeGraph) -> Result<()> {
-        let start = Instant::now();
-
-        // Build adjacency matrix
-        let entities: Vec<EntityId> = graph.entities().map(|e| e.id.clone()).collect();
-        let n = entities.len();
-
-        if n == 0 {
-            return Ok(());
-        }
-
-        // Initialize scores
-        let initial_score = 1.0 / n as f64;
-        for entity_id in &entities {
-            self.scores.insert(entity_id.clone(), initial_score);
-        }
-
-        // Power iteration
-        for iteration in 0..self.max_iterations {
-            let mut new_scores = HashMap::new();
-            let mut max_diff: f64 = 0.0;
-
-            for entity_id in &entities {
-                let mut score = (1.0 - self.damping_factor) / n as f64;
-
-                // Sum contributions from incoming links
-                for other_entity in &entities {
-                    if let Some(weight) = self.get_edge_weight(other_entity, entity_id, graph) {
-                        let other_score = self
-                            .scores
-                            .get(other_entity)
-                            .map(|s| *s.value())
-                            .unwrap_or(initial_score);
-                        let out_degree = self.get_out_degree(other_entity, graph);
-
-                        if out_degree > 0.0 {
-                            score += self.damping_factor * other_score * weight / out_degree;
-                        }
-                    }
-                }
-
-                let old_score = self
-                    .scores
-                    .get(entity_id)
-                    .map(|s| *s.value())
-                    .unwrap_or(initial_score);
-                let diff = (score - old_score).abs();
-                max_diff = max_diff.max(diff);
-
-                new_scores.insert(entity_id.clone(), score);
-            }
-
-            // Update scores
-            for (entity_id, score) in new_scores {
-                self.scores.insert(entity_id, score);
-            }
-
-            // Check convergence
-            if max_diff < self.tolerance {
-                println!(
-                    "🎯 PageRank converged after {} iterations (diff: {:.6})",
-                    iteration + 1,
-                    max_diff
-                );
-                break;
-            }
-        }
-
-        let duration = start.elapsed();
-        println!("🔄 Full PageRank recomputation completed in {duration:?} for {n} entities");
-
-        Ok(())
-    }
-
-    /// Perform localized PageRank computation for a subset of entities
-    async fn localized_pagerank(
-        &self,
-        entities: &HashSet<EntityId>,
-        graph: &KnowledgeGraph,
-    ) -> Result<()> {
-        let entity_vec: Vec<EntityId> = entities.iter().cloned().collect();
-        let n = entity_vec.len();
-
-        if n == 0 {
-            return Ok(());
-        }
-
-        // Localized power iteration
-        for _iteration in 0..self.max_iterations {
-            let mut max_diff: f64 = 0.0;
-
-            for entity_id in &entity_vec {
-                let mut score = (1.0 - self.damping_factor) / n as f64;
-
-                // Only consider links within the subset for localized computation
-                for other_entity in &entity_vec {
-                    if let Some(weight) = self.get_edge_weight(other_entity, entity_id, graph) {
-                        let other_score = self
-                            .scores
-                            .get(other_entity)
-                            .map(|s| *s.value())
-                            .unwrap_or(1.0 / n as f64);
-                        let out_degree =
-                            self.get_localized_out_degree(other_entity, entities, graph);
-
-                        if out_degree > 0.0 {
-                            score += self.damping_factor * other_score * weight / out_degree;
-                        }
-                    }
-                }
-
-                let old_score = self
-                    .scores
-                    .get(entity_id)
-                    .map(|s| *s.value())
-                    .unwrap_or(1.0 / n as f64);
-                let diff = (score - old_score).abs();
-                max_diff = max_diff.max(diff);
-
-                self.scores.insert(entity_id.clone(), score);
-            }
-
-            // Check convergence
-            if max_diff < self.tolerance {
-                break;
-            }
-        }
-
-        Ok(())
-    }
-
-    fn get_edge_weight(
-        &self,
-        from: &EntityId,
-        to: &EntityId,
-        graph: &KnowledgeGraph,
-    ) -> Option<f64> {
-        // Check if there's a relationship between entities
-        for (neighbor, relationship) in graph.get_neighbors(from) {
-            if neighbor.id == *to {
-                return Some(relationship.confidence as f64);
-            }
-        }
-        None
-    }
-
-    fn get_out_degree(&self, entity_id: &EntityId, graph: &KnowledgeGraph) -> f64 {
-        graph
-            .get_neighbors(entity_id)
-            .iter()
-            .map(|(_, rel)| rel.confidence as f64)
-            .sum()
-    }
-
-    fn get_localized_out_degree(
-        &self,
-        entity_id: &EntityId,
-        subset: &HashSet<EntityId>,
-        graph: &KnowledgeGraph,
-    ) -> f64 {
-        graph
-            .get_neighbors(entity_id)
-            .iter()
-            .filter(|(neighbor, _)| subset.contains(&neighbor.id))
-            .map(|(_, rel)| rel.confidence as f64)
-            .sum()
-    }
-
-    /// Get PageRank score for an entity
-    pub fn get_score(&self, entity_id: &EntityId) -> Option<f64> {
-        self.scores.get(entity_id).map(|s| *s.value())
-    }
-
-    /// Get top-k entities by PageRank score
-    pub fn get_top_entities(&self, k: usize) -> Vec<(EntityId, f64)> {
-        let mut entities: Vec<(EntityId, f64)> = self
-            .scores
-            .iter()
-            .map(|entry| (entry.key().clone(), *entry.value()))
-            .collect();
-
-        entities.sort_by(|a, b| b.1.partial_cmp(&a.1).unwrap_or(std::cmp::Ordering::Equal));
-        entities.truncate(k);
-        entities
-    }
-
-    /// Record a graph change for incremental updates
-    pub fn record_change(&self, _entity_id: EntityId) {
-        *self.pending_changes.write() += 1;
-    }
-}
-
-// ============================================================================
-// Batch Processing System
-// ============================================================================
-
-/// High-throughput batch processor for incremental updates
-#[cfg(feature = "incremental")]
-pub struct BatchProcessor {
-    batch_size: usize,
-    max_wait_time: Duration,
-    pending_batches: DashMap<String, PendingBatch>,
-    processing_semaphore: Semaphore,
-    metrics: RwLock<BatchMetrics>,
-}
-
-#[cfg(feature = "incremental")]
-#[derive(Debug, Clone)]
-struct PendingBatch {
-    changes: Vec<ChangeRecord>,
-    created_at: Instant,
-    batch_id: String,
-}
-
-/// Batch metrics for monitoring
-#[derive(Debug, Clone)]
-pub struct BatchMetrics {
-    /// Total number of batches processed
-    pub total_batches_processed: u64,
-    /// Total number of changes processed across all batches
-    pub total_changes_processed: u64,
-    /// Average size of batches
-    pub average_batch_size: f64,
-    /// Average time to process a batch
-    pub average_processing_time: Duration,
-    /// Throughput in changes per second
-    pub throughput_per_second: f64,
-    /// Timestamp of last batch processed
-    pub last_batch_processed: Option<DateTime<Utc>>,
-}
-
-#[cfg(feature = "incremental")]
-impl BatchProcessor {
-    /// Creates a new batch processor with specified configuration
-    pub fn new(batch_size: usize, max_wait_time: Duration, max_concurrent_batches: usize) -> Self {
-        Self {
-            batch_size,
-            max_wait_time,
-            pending_batches: DashMap::new(),
-            processing_semaphore: Semaphore::new(max_concurrent_batches),
-            metrics: RwLock::new(BatchMetrics {
-                total_batches_processed: 0,
-                total_changes_processed: 0,
-                average_batch_size: 0.0,
-                average_processing_time: Duration::from_millis(0),
-                throughput_per_second: 0.0,
-                last_batch_processed: None,
-            }),
-        }
-    }
-
-    /// Adds a change to be processed in batches
-    pub async fn add_change(&self, change: ChangeRecord) -> Result<String> {
-        let batch_key = self.get_batch_key(&change);
-
-        // Capture the batch snapshot and id under the entry RefMut, then
-        // drop the RefMut before touching pending_batches again. DashMap
-        // remove on the same key while a RefMut is alive deadlocks on the
-        // shard write lock the RefMut holds.
-        let (batch_id, batch_to_flush) = {
-            let mut entry = self
-                .pending_batches
-                .entry(batch_key.clone())
-                .or_insert_with(|| PendingBatch {
-                    changes: Vec::new(),
-                    created_at: Instant::now(),
-                    batch_id: format!("batch_{}", Uuid::new_v4()),
-                });
-
-            entry.changes.push(change);
-            let should_process = entry.changes.len() >= self.batch_size
-                || entry.created_at.elapsed() > self.max_wait_time;
-
-            let batch_id = entry.batch_id.clone();
-            let batch_to_flush = should_process.then(|| entry.clone());
-            (batch_id, batch_to_flush)
-        };
-
-        if let Some(batch) = batch_to_flush {
-            self.pending_batches.remove(&batch_key);
-
-            let processor = Arc::new(self.clone());
-            tokio::spawn(async move {
-                if let Err(e) = processor.process_batch(batch).await {
-                    eprintln!("Batch processing error: {e}");
-                }
-            });
-        }
-
-        Ok(batch_id)
-    }
-
-    async fn process_batch(&self, batch: PendingBatch) -> Result<()> {
-        let _permit = self.processing_semaphore.acquire().await.map_err(|_| {
-            GraphRAGError::IncrementalUpdate {
-                message: "Failed to acquire processing permit".to_string(),
-            }
-        })?;
-
-        let start = Instant::now();
-
-        // Group changes by type for optimized processing
-        let mut entity_changes = Vec::new();
-        let mut relationship_changes = Vec::new();
-        let mut embedding_changes = Vec::new();
-
-        for change in &batch.changes {
-            match &change.change_type {
-                ChangeType::EntityAdded | ChangeType::EntityUpdated | ChangeType::EntityRemoved => {
-                    entity_changes.push(change);
-                },
-                ChangeType::RelationshipAdded
-                | ChangeType::RelationshipUpdated
-                | ChangeType::RelationshipRemoved => {
-                    relationship_changes.push(change);
-                },
-                ChangeType::EmbeddingAdded
-                | ChangeType::EmbeddingUpdated
-                | ChangeType::EmbeddingRemoved => {
-                    embedding_changes.push(change);
-                },
-                _ => {},
-            }
-        }
-
-        // Process each type of change optimally
-        self.process_entity_changes(&entity_changes).await?;
-        self.process_relationship_changes(&relationship_changes)
-            .await?;
-        self.process_embedding_changes(&embedding_changes).await?;
-
-        let processing_time = start.elapsed();
-
-        // Update metrics
-        self.update_metrics(&batch, processing_time).await;
-
-        println!(
-            "🚀 Processed batch {} with {} changes in {:?}",
-            batch.batch_id,
-            batch.changes.len(),
-            processing_time
-        );
-
-        Ok(())
-    }
-
-    async fn process_entity_changes(&self, _changes: &[&ChangeRecord]) -> Result<()> {
-        // Implementation would go here - process entity changes efficiently
-        Ok(())
-    }
-
-    async fn process_relationship_changes(&self, _changes: &[&ChangeRecord]) -> Result<()> {
-        // Implementation would go here - process relationship changes efficiently
-        Ok(())
-    }
-
-    async fn process_embedding_changes(&self, _changes: &[&ChangeRecord]) -> Result<()> {
-        // Implementation would go here - process embedding changes efficiently
-        Ok(())
-    }
-
-    fn get_batch_key(&self, change: &ChangeRecord) -> String {
-        // Group changes by entity or document for batching efficiency
-        match (&change.entity_id, &change.document_id) {
-            (Some(entity_id), _) => format!("entity:{entity_id}"),
-            (None, Some(doc_id)) => format!("document:{doc_id}"),
-            _ => "global".to_string(),
-        }
-    }
-
-    async fn update_metrics(&self, batch: &PendingBatch, processing_time: Duration) {
-        let mut metrics = self.metrics.write();
-
-        metrics.total_batches_processed += 1;
-        metrics.total_changes_processed += batch.changes.len() as u64;
-
-        // Update running averages
-        let total_batches = metrics.total_batches_processed as f64;
-        metrics.average_batch_size = (metrics.average_batch_size * (total_batches - 1.0)
-            + batch.changes.len() as f64)
-            / total_batches;
-
-        let prev_avg_ms = metrics.average_processing_time.as_millis() as f64;
-        let new_avg_ms = (prev_avg_ms * (total_batches - 1.0) + processing_time.as_millis() as f64)
-            / total_batches;
-        metrics.average_processing_time = Duration::from_millis(new_avg_ms as u64);
-
-        // Calculate throughput
-        if processing_time.as_secs_f64() > 0.0 {
-            metrics.throughput_per_second =
-                batch.changes.len() as f64 / processing_time.as_secs_f64();
-        }
-
-        metrics.last_batch_processed = Some(Utc::now());
-    }
-
-    /// Gets the current batch processing metrics
-    pub fn get_metrics(&self) -> BatchMetrics {
-        self.metrics.read().clone()
-    }
-}
-
-// Clone impl for BatchProcessor (required for Arc usage)
-#[cfg(feature = "incremental")]
-impl Clone for BatchProcessor {
-    fn clone(&self) -> Self {
-        Self {
-            batch_size: self.batch_size,
-            max_wait_time: self.max_wait_time,
-            pending_batches: DashMap::new(), // New instance starts empty
-            processing_semaphore: Semaphore::new(self.processing_semaphore.available_permits()),
-            metrics: RwLock::new(self.get_metrics()),
-        }
-    }
-}
-
-// ============================================================================
-// Error Extensions
-// ============================================================================
-
-impl GraphRAGError {
-    /// Creates a conflict resolution error
-    pub fn conflict_resolution(message: String) -> Self {
-        GraphRAGError::GraphConstruction { message }
-    }
-
-    /// Creates an incremental update error
-    pub fn incremental_update(message: String) -> Self {
-        GraphRAGError::GraphConstruction { message }
-    }
-}
-
-// ============================================================================
-// Production-Ready IncrementalGraphStore Implementation
-// ============================================================================
-
-/// Production implementation of IncrementalGraphStore with full ACID guarantees
-#[cfg(feature = "incremental")]
-pub struct ProductionGraphStore {
-    graph: Arc<RwLock<KnowledgeGraph>>,
-    transactions: DashMap<TransactionId, Transaction>,
-    change_log: DashMap<UpdateId, ChangeRecord>,
-    rollback_data: DashMap<UpdateId, RollbackData>,
-    conflict_resolver: Arc<ConflictResolver>,
-    cache_invalidation: Arc<SelectiveInvalidation>,
-    monitor: Arc<UpdateMonitor>,
-    // Held for batch-update wiring; current code path goes directly through `apply_change`.
-    #[allow(dead_code)]
-    batch_processor: Arc<BatchProcessor>,
-    incremental_pagerank: Arc<IncrementalPageRank>,
-    event_publisher: broadcast::Sender<ChangeEvent>,
-    // Captured at construction for future runtime tuning hooks.
-    #[allow(dead_code)]
-    config: IncrementalConfig,
-}
-
-/// Transaction state for ACID operations
-#[cfg(feature = "incremental")]
-#[derive(Debug, Clone)]
-struct Transaction {
-    // Tx id is also the DashMap key; field is for debug/logging.
-    #[allow(dead_code)]
-    id: TransactionId,
-    changes: Vec<ChangeRecord>,
-    status: TransactionStatus,
-    // Recorded for transaction-age metrics; reader pending.
-    #[allow(dead_code)]
-    created_at: DateTime<Utc>,
-    // Reserved for per-transaction isolation policy enforcement.
-    #[allow(dead_code)]
-    isolation_level: IsolationLevel,
-}
-
-#[cfg(feature = "incremental")]
-#[derive(Debug, Clone, PartialEq)]
-#[allow(dead_code)] // Variants reserved for two-phase commit / abort flows; ABI-stable surface.
-enum TransactionStatus {
-    Active,
-    Preparing,
-    Committed,
-    Aborted,
-}
-
-#[cfg(feature = "incremental")]
-#[derive(Debug, Clone)]
-#[allow(dead_code)] // Variants reserved for isolation-level configuration; ABI-stable surface.
-enum IsolationLevel {
-    ReadUncommitted,
-    ReadCommitted,
-    RepeatableRead,
-    Serializable,
-}
-
-/// Change event for monitoring and debugging
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChangeEvent {
-    /// Unique identifier for the event
-    pub event_id: UpdateId,
-    /// Type of change event
-    pub event_type: ChangeEventType,
-    /// Optional entity ID associated with the event
-    pub entity_id: Option<EntityId>,
-    /// When the event occurred
-    pub timestamp: DateTime<Utc>,
-    /// Additional metadata about the event
-    pub metadata: HashMap<String, String>,
-}
-
-/// Types of change events that can be published
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ChangeEventType {
-    /// An entity was upserted
-    EntityUpserted,
-    /// An entity was deleted
-    EntityDeleted,
-    /// A relationship was upserted
-    RelationshipUpserted,
-    /// A relationship was deleted
-    RelationshipDeleted,
-    /// An embedding was updated
-    EmbeddingUpdated,
-    /// A transaction was started
-    TransactionStarted,
-    /// A transaction was committed
-    TransactionCommitted,
-    /// A transaction was rolled back
-    TransactionRolledBack,
-    /// A conflict was resolved
-    ConflictResolved,
-    /// Cache was invalidated
-    CacheInvalidated,
-    /// A batch was processed
-    BatchProcessed,
-}
-
-#[cfg(feature = "incremental")]
-impl ProductionGraphStore {
-    /// Creates a new production-grade graph store with full ACID guarantees
-    pub fn new(
-        graph: KnowledgeGraph,
-        config: IncrementalConfig,
-        conflict_resolver: ConflictResolver,
-    ) -> Self {
-        let (event_tx, _) = broadcast::channel(1000);
-
-        Self {
-            graph: Arc::new(RwLock::new(graph)),
-            transactions: DashMap::new(),
-            change_log: DashMap::new(),
-            rollback_data: DashMap::new(),
-            conflict_resolver: Arc::new(conflict_resolver),
-            cache_invalidation: Arc::new(SelectiveInvalidation::new()),
-            monitor: Arc::new(UpdateMonitor::new()),
-            batch_processor: Arc::new(BatchProcessor::new(
-                config.batch_size,
-                Duration::from_millis(100),
-                config.max_concurrent_operations,
-            )),
-            incremental_pagerank: Arc::new(IncrementalPageRank::new(0.85, 1e-6, 100)),
-            event_publisher: event_tx,
-            config,
-        }
-    }
-
-    /// Subscribes to change events for monitoring
-    pub fn subscribe_events(&self) -> broadcast::Receiver<ChangeEvent> {
-        self.event_publisher.subscribe()
-    }
-
-    async fn publish_event(&self, event: ChangeEvent) {
-        let _ = self.event_publisher.send(event);
-    }
-
-    fn create_change_record(
-        &self,
-        change_type: ChangeType,
-        operation: Operation,
-        change_data: ChangeData,
-        entity_id: Option<EntityId>,
-        document_id: Option<DocumentId>,
-    ) -> ChangeRecord {
-        ChangeRecord {
-            change_id: UpdateId::new(),
-            timestamp: Utc::now(),
-            change_type,
-            entity_id,
-            document_id,
-            operation,
-            data: change_data,
-            metadata: HashMap::new(),
-        }
-    }
-
-    async fn apply_change_with_conflict_resolution(
-        &self,
-        change: ChangeRecord,
-    ) -> Result<UpdateId> {
-        let operation_id = self.monitor.start_operation("apply_change");
-        // Capture change_id for the return value: change_log is keyed by
-        // change.change_id, not by the monitor's operation_id, so callers
-        // that look the returned id up in change_log would otherwise miss.
-        let change_id = change.change_id.clone();
-
-        // Check for conflicts
-        if let Some(conflict) = self.detect_conflict(&change)? {
-            let resolution = self.conflict_resolver.resolve_conflict(&conflict).await?;
-
-            // Apply resolved change
-            let resolved_change = ChangeRecord {
-                data: resolution.resolved_data,
-                metadata: resolution.metadata,
-                ..change
-            };
-
-            self.apply_change_internal(resolved_change).await?;
-
-            // Publish conflict resolution event
-            self.publish_event(ChangeEvent {
-                event_id: UpdateId::new(),
-                event_type: ChangeEventType::ConflictResolved,
-                entity_id: conflict.existing_data.get_entity_id(),
-                timestamp: Utc::now(),
-                metadata: HashMap::new(),
-            })
-            .await;
-        } else {
-            self.apply_change_internal(change).await?;
-        }
-
-        self.monitor
-            .complete_operation(&operation_id, true, None, 1, 0);
-        Ok(change_id)
-    }
-
-    fn detect_conflict(&self, change: &ChangeRecord) -> Result<Option<Conflict>> {
-        match &change.data {
-            ChangeData::Entity(entity) => {
-                let graph = self.graph.read();
-                if let Some(existing) = graph.get_entity(&entity.id) {
-                    if existing.name != entity.name || existing.entity_type != entity.entity_type {
-                        return Ok(Some(Conflict {
-                            conflict_id: UpdateId::new(),
-                            conflict_type: ConflictType::EntityExists,
-                            existing_data: ChangeData::Entity(existing.clone()),
-                            new_data: change.data.clone(),
-                            resolution: None,
-                        }));
-                    }
-                }
-            },
-            ChangeData::Relationship(relationship) => {
-                let graph = self.graph.read();
-                for existing_rel in graph.get_all_relationships() {
-                    if existing_rel.source == relationship.source
-                        && existing_rel.target == relationship.target
-                        && existing_rel.relation_type == relationship.relation_type
-                    {
-                        return Ok(Some(Conflict {
-                            conflict_id: UpdateId::new(),
-                            conflict_type: ConflictType::RelationshipExists,
-                            existing_data: ChangeData::Relationship(existing_rel.clone()),
-                            new_data: change.data.clone(),
-                            resolution: None,
-                        }));
-                    }
-                }
-            },
-            _ => {},
-        }
-
-        Ok(None)
-    }
-
-    async fn apply_change_internal(&self, change: ChangeRecord) -> Result<()> {
-        let change_id = change.change_id.clone();
-
-        // Create rollback data first
-        let rollback_data = {
-            let graph = self.graph.read();
-            self.create_rollback_data(&change, &graph)?
-        };
-
-        self.rollback_data.insert(change_id.clone(), rollback_data);
-
-        // Apply the change
-        {
-            let mut graph = self.graph.write();
-            match &change.data {
-                ChangeData::Entity(entity) => {
-                    match change.operation {
-                        Operation::Insert | Operation::Upsert => {
-                            graph.add_entity(entity.clone())?;
-                            self.incremental_pagerank.record_change(entity.id.clone());
-                        },
-                        Operation::Delete => {
-                            // Remove entity and its relationships
-                            // Implementation would go here
-                        },
-                        _ => {},
-                    }
-                },
-                ChangeData::Relationship(relationship) => {
-                    match change.operation {
-                        Operation::Insert | Operation::Upsert => {
-                            graph.add_relationship(relationship.clone())?;
-                            self.incremental_pagerank
-                                .record_change(relationship.source.clone());
-                            self.incremental_pagerank
-                                .record_change(relationship.target.clone());
-                        },
-                        Operation::Delete => {
-                            // Remove relationship
-                            // Implementation would go here
-                        },
-                        _ => {},
-                    }
-                },
-                ChangeData::Embedding {
-                    entity_id,
-                    embedding,
-                } => {
-                    if let Some(entity) = graph.get_entity_mut(entity_id) {
-                        entity.embedding = Some(embedding.clone());
-                    }
-                },
-                _ => {},
-            }
-        }
-
-        // Record change in log
-        self.change_log.insert(change_id, change);
-
-        Ok(())
-    }
-
-    fn create_rollback_data(
-        &self,
-        change: &ChangeRecord,
-        graph: &KnowledgeGraph,
-    ) -> Result<RollbackData> {
-        let mut previous_entities = Vec::new();
-        let mut previous_relationships = Vec::new();
-
-        match &change.data {
-            ChangeData::Entity(entity) => {
-                if let Some(existing) = graph.get_entity(&entity.id) {
-                    previous_entities.push(existing.clone());
-                }
-            },
-            ChangeData::Relationship(relationship) => {
-                // Store existing relationships that might be affected
-                for rel in graph.get_all_relationships() {
-                    if rel.source == relationship.source && rel.target == relationship.target {
-                        previous_relationships.push(rel.clone());
-                    }
-                }
-            },
-            _ => {},
-        }
-
-        Ok(RollbackData {
-            previous_entities,
-            previous_relationships,
-            affected_caches: vec![], // Will be populated by cache invalidation system
-        })
-    }
-}
-
-#[cfg(feature = "incremental")]
-#[async_trait::async_trait]
-impl IncrementalGraphStore for ProductionGraphStore {
-    type Error = GraphRAGError;
-
-    async fn upsert_entity(&mut self, entity: Entity) -> Result<UpdateId> {
-        let change = self.create_change_record(
-            ChangeType::EntityAdded,
-            Operation::Upsert,
-            ChangeData::Entity(entity.clone()),
-            Some(entity.id.clone()),
-            None,
-        );
-
-        let update_id = self.apply_change_with_conflict_resolution(change).await?;
-
-        // Trigger cache invalidation
-        let changes = vec![self.change_log.get(&update_id).unwrap().clone()];
-        let _invalidation_strategies = self.cache_invalidation.invalidate_for_changes(&changes);
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::EntityUpserted,
-            entity_id: Some(entity.id),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        })
-        .await;
-
-        Ok(update_id)
-    }
-
-    async fn upsert_relationship(&mut self, relationship: Relationship) -> Result<UpdateId> {
-        let change = self.create_change_record(
-            ChangeType::RelationshipAdded,
-            Operation::Upsert,
-            ChangeData::Relationship(relationship.clone()),
-            None,
-            None,
-        );
-
-        let update_id = self.apply_change_with_conflict_resolution(change).await?;
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::RelationshipUpserted,
-            entity_id: Some(relationship.source),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        })
-        .await;
-
-        Ok(update_id)
-    }
-
-    async fn delete_entity(&mut self, entity_id: &EntityId) -> Result<UpdateId> {
-        // Implementation for entity deletion
-        let update_id = UpdateId::new();
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::EntityDeleted,
-            entity_id: Some(entity_id.clone()),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        })
-        .await;
-
-        Ok(update_id)
-    }
-
-    async fn delete_relationship(
-        &mut self,
-        source: &EntityId,
-        _target: &EntityId,
-        _relation_type: &str,
-    ) -> Result<UpdateId> {
-        // Implementation for relationship deletion
-        let update_id = UpdateId::new();
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::RelationshipDeleted,
-            entity_id: Some(source.clone()),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        })
-        .await;
-
-        Ok(update_id)
-    }
-
-    async fn apply_delta(&mut self, delta: GraphDelta) -> Result<UpdateId> {
-        let tx_id = self.begin_transaction().await?;
-
-        for change in delta.changes {
-            self.apply_change_with_conflict_resolution(change).await?;
-        }
-
-        self.commit_transaction(tx_id).await?;
-        Ok(delta.delta_id)
-    }
-
-    async fn rollback_delta(&mut self, _delta_id: &UpdateId) -> Result<()> {
-        // Implementation for delta rollback
-        Ok(())
-    }
-
-    async fn get_change_log(&self, since: Option<DateTime<Utc>>) -> Result<Vec<ChangeRecord>> {
-        let changes: Vec<ChangeRecord> = self
-            .change_log
-            .iter()
-            .filter_map(|entry| {
-                let change = entry.value();
-                if let Some(since_time) = since {
-                    if change.timestamp >= since_time {
-                        Some(change.clone())
-                    } else {
-                        None
-                    }
-                } else {
-                    Some(change.clone())
-                }
-            })
-            .collect();
-
-        Ok(changes)
-    }
-
-    async fn begin_transaction(&mut self) -> Result<TransactionId> {
-        let tx_id = TransactionId::new();
-        let transaction = Transaction {
-            id: tx_id.clone(),
-            changes: Vec::new(),
-            status: TransactionStatus::Active,
-            created_at: Utc::now(),
-            isolation_level: IsolationLevel::ReadCommitted,
-        };
-
-        self.transactions.insert(tx_id.clone(), transaction);
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::TransactionStarted,
-            entity_id: None,
-            timestamp: Utc::now(),
-            metadata: [("transaction_id".to_string(), tx_id.to_string())]
-                .into_iter()
-                .collect(),
-        })
-        .await;
-
-        Ok(tx_id)
-    }
-
-    async fn commit_transaction(&mut self, tx_id: TransactionId) -> Result<()> {
-        if let Some((_, mut tx)) = self.transactions.remove(&tx_id) {
-            tx.status = TransactionStatus::Committed;
-
-            // Publish event
-            self.publish_event(ChangeEvent {
-                event_id: UpdateId::new(),
-                event_type: ChangeEventType::TransactionCommitted,
-                entity_id: None,
-                timestamp: Utc::now(),
-                metadata: [("transaction_id".to_string(), tx_id.to_string())]
-                    .into_iter()
-                    .collect(),
-            })
-            .await;
-
-            Ok(())
-        } else {
-            Err(GraphRAGError::IncrementalUpdate {
-                message: format!("Transaction {tx_id} not found"),
-            })
-        }
-    }
-
-    async fn rollback_transaction(&mut self, tx_id: TransactionId) -> Result<()> {
-        if let Some((_, mut tx)) = self.transactions.remove(&tx_id) {
-            tx.status = TransactionStatus::Aborted;
-
-            // Rollback all changes in this transaction
-            for _change in &tx.changes {
-                // Implementation for rollback
-            }
-
-            // Publish event
-            self.publish_event(ChangeEvent {
-                event_id: UpdateId::new(),
-                event_type: ChangeEventType::TransactionRolledBack,
-                entity_id: None,
-                timestamp: Utc::now(),
-                metadata: [("transaction_id".to_string(), tx_id.to_string())]
-                    .into_iter()
-                    .collect(),
-            })
-            .await;
-
-            Ok(())
-        } else {
-            Err(GraphRAGError::IncrementalUpdate {
-                message: format!("Transaction {tx_id} not found"),
-            })
-        }
-    }
-
-    async fn batch_upsert_entities(
-        &mut self,
-        entities: Vec<Entity>,
-        _strategy: ConflictStrategy,
-    ) -> Result<Vec<UpdateId>> {
-        let mut update_ids = Vec::new();
-
-        for entity in entities {
-            let update_id = self.upsert_entity(entity).await?;
-            update_ids.push(update_id);
-        }
-
-        Ok(update_ids)
-    }
-
-    async fn batch_upsert_relationships(
-        &mut self,
-        relationships: Vec<Relationship>,
-        _strategy: ConflictStrategy,
-    ) -> Result<Vec<UpdateId>> {
-        let mut update_ids = Vec::new();
-
-        for relationship in relationships {
-            let update_id = self.upsert_relationship(relationship).await?;
-            update_ids.push(update_id);
-        }
-
-        Ok(update_ids)
-    }
-
-    async fn update_entity_embedding(
-        &mut self,
-        entity_id: &EntityId,
-        embedding: Vec<f32>,
-    ) -> Result<UpdateId> {
-        let change = self.create_change_record(
-            ChangeType::EmbeddingUpdated,
-            Operation::Update,
-            ChangeData::Embedding {
-                entity_id: entity_id.clone(),
-                embedding,
-            },
-            Some(entity_id.clone()),
-            None,
-        );
-
-        let update_id = self.apply_change_with_conflict_resolution(change).await?;
-
-        // Publish event
-        self.publish_event(ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::EmbeddingUpdated,
-            entity_id: Some(entity_id.clone()),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        })
-        .await;
-
-        Ok(update_id)
-    }
-
-    async fn bulk_update_embeddings(
-        &mut self,
-        updates: Vec<(EntityId, Vec<f32>)>,
-    ) -> Result<Vec<UpdateId>> {
-        let mut update_ids = Vec::new();
-
-        for (entity_id, embedding) in updates {
-            let update_id = self.update_entity_embedding(&entity_id, embedding).await?;
-            update_ids.push(update_id);
-        }
-
-        Ok(update_ids)
-    }
-
-    async fn get_pending_transactions(&self) -> Result<Vec<TransactionId>> {
-        let pending: Vec<TransactionId> = self
-            .transactions
-            .iter()
-            .filter(|entry| entry.value().status == TransactionStatus::Active)
-            .map(|entry| entry.key().clone())
-            .collect();
-
-        Ok(pending)
-    }
-
-    async fn get_graph_statistics(&self) -> Result<GraphStatistics> {
-        let graph = self.graph.read();
-        let entities: Vec<_> = graph.entities().collect();
-        let relationships = graph.get_all_relationships();
-
-        let node_count = entities.len();
-        let edge_count = relationships.len();
-
-        // Calculate average degree
-        let total_degree: usize = entities
-            .iter()
-            .map(|entity| graph.get_neighbors(&entity.id).len())
-            .sum();
-
-        let average_degree = if node_count > 0 {
-            total_degree as f64 / node_count as f64
-        } else {
-            0.0
-        };
-
-        // Find max degree
-        let max_degree = entities
-            .iter()
-            .map(|entity| graph.get_neighbors(&entity.id).len())
-            .max()
-            .unwrap_or(0);
-
-        Ok(GraphStatistics {
-            node_count,
-            edge_count,
-            average_degree,
-            max_degree,
-            connected_components: 1,     // Simplified for now
-            clustering_coefficient: 0.0, // Would need complex calculation
-            last_updated: Utc::now(),
-        })
-    }
-
-    async fn validate_consistency(&self) -> Result<ConsistencyReport> {
-        let graph = self.graph.read();
-        let mut orphaned_entities = Vec::new();
-        let mut broken_relationships = Vec::new();
-        let mut missing_embeddings = Vec::new();
-
-        // Check for orphaned entities (entities with no relationships)
-        for entity in graph.entities() {
-            let neighbors = graph.get_neighbors(&entity.id);
-            if neighbors.is_empty() {
-                orphaned_entities.push(entity.id.clone());
-            }
-
-            // Check for missing embeddings
-            if entity.embedding.is_none() {
-                missing_embeddings.push(entity.id.clone());
-            }
-        }
-
-        // Check for broken relationships (references to non-existent entities)
-        for relationship in graph.get_all_relationships() {
-            if graph.get_entity(&relationship.source).is_none()
-                || graph.get_entity(&relationship.target).is_none()
-            {
-                broken_relationships.push((
-                    relationship.source.clone(),
-                    relationship.target.clone(),
-                    relationship.relation_type.clone(),
-                ));
-            }
-        }
-
-        let issues_found =
-            orphaned_entities.len() + broken_relationships.len() + missing_embeddings.len();
-
-        Ok(ConsistencyReport {
-            is_consistent: issues_found == 0,
-            orphaned_entities,
-            broken_relationships,
-            missing_embeddings,
-            validation_time: Utc::now(),
-            issues_found,
-        })
-    }
-}
-
-// Helper trait for extracting entity ID from ChangeData
-#[allow(dead_code)] // Used only when feature = "incremental" is enabled.
-trait ChangeDataExt {
-    fn get_entity_id(&self) -> Option<EntityId>;
-}
-
-impl ChangeDataExt for ChangeData {
-    fn get_entity_id(&self) -> Option<EntityId> {
-        match self {
-            ChangeData::Entity(entity) => Some(entity.id.clone()),
-            ChangeData::Embedding { entity_id, .. } => Some(entity_id.clone()),
-            _ => None,
-        }
-    }
-}
-
-// Re-export for backward compatibility - removing to avoid duplicate definition
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_update_id_generation() {
-        let id1 = UpdateId::new();
-        let id2 = UpdateId::new();
-        assert_ne!(id1.as_str(), id2.as_str());
-    }
-
-    #[test]
-    fn test_transaction_id_generation() {
-        let tx1 = TransactionId::new();
-        let tx2 = TransactionId::new();
-        assert_ne!(tx1.as_str(), tx2.as_str());
-    }
-
-    #[test]
-    fn test_change_record_creation() {
-        let entity = Entity::new(
-            EntityId::new("test".to_string()),
-            "Test Entity".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let config = IncrementalConfig::default();
-        let graph = KnowledgeGraph::new();
-        let manager = IncrementalGraphManager::new(graph, config);
-
-        let change = manager.create_change_record(
-            ChangeType::EntityAdded,
-            Operation::Insert,
-            ChangeData::Entity(entity.clone()),
-            Some(entity.id.clone()),
-            None,
-        );
-
-        assert_eq!(change.change_type, ChangeType::EntityAdded);
-        assert_eq!(change.operation, Operation::Insert);
-        assert_eq!(change.entity_id, Some(entity.id));
-    }
-
-    #[test]
-    fn test_conflict_resolver_creation() {
-        let resolver = ConflictResolver::new(ConflictStrategy::KeepExisting);
-        assert!(matches!(resolver.strategy, ConflictStrategy::KeepExisting));
-    }
-
-    #[test]
-    fn test_incremental_config_default() {
-        let config = IncrementalConfig::default();
-        assert_eq!(config.max_change_log_size, 10000);
-        assert_eq!(config.batch_size, 100);
-        assert!(config.enable_monitoring);
-    }
-
-    #[test]
-    fn test_statistics_creation() {
-        let stats = IncrementalStatistics::empty();
-        assert_eq!(stats.total_updates, 0);
-        assert_eq!(stats.entities_added, 0);
-        assert_eq!(stats.average_update_time_ms, 0.0);
-    }
-
-    #[tokio::test]
-    async fn test_basic_entity_upsert() {
-        let config = IncrementalConfig::default();
-        let graph = KnowledgeGraph::new();
-        let mut manager = IncrementalGraphManager::new(graph, config);
-
-        let entity = Entity::new(
-            EntityId::new("test_entity".to_string()),
-            "Test Entity".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let update_id = manager.basic_upsert_entity(entity).unwrap();
-        assert!(!update_id.as_str().is_empty());
-
-        let stats = manager.get_statistics();
-        assert_eq!(stats.entities_added, 1);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_creation() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let store = ProductionGraphStore::new(graph, config, resolver);
-        let _receiver = store.subscribe_events();
-        // If we reached here, subscription succeeded; no further assertion needed.
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_entity_upsert() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let mut store = ProductionGraphStore::new(graph, config, resolver);
-
-        let entity = Entity::new(
-            EntityId::new("test_entity".to_string()),
-            "Test Entity".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let update_id = store.upsert_entity(entity).await.unwrap();
-        assert!(!update_id.as_str().is_empty());
-
-        let stats = store.get_graph_statistics().await.unwrap();
-        assert_eq!(stats.node_count, 1);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_relationship_upsert() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let mut store = ProductionGraphStore::new(graph, config, resolver);
-
-        // Add entities first
-        let entity1 = Entity::new(
-            EntityId::new("entity1".to_string()),
-            "Entity 1".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let entity2 = Entity::new(
-            EntityId::new("entity2".to_string()),
-            "Entity 2".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        store.upsert_entity(entity1.clone()).await.unwrap();
-        store.upsert_entity(entity2.clone()).await.unwrap();
-
-        let relationship = Relationship::new(entity1.id, entity2.id, "KNOWS".to_string(), 0.8);
-
-        let update_id = store.upsert_relationship(relationship).await.unwrap();
-        assert!(!update_id.as_str().is_empty());
-
-        let stats = store.get_graph_statistics().await.unwrap();
-        assert_eq!(stats.edge_count, 1);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_transactions() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let mut store = ProductionGraphStore::new(graph, config, resolver);
-
-        let tx_id = store.begin_transaction().await.unwrap();
-        assert!(!tx_id.as_str().is_empty());
-
-        let pending = store.get_pending_transactions().await.unwrap();
-        assert_eq!(pending.len(), 1);
-        assert_eq!(pending[0], tx_id);
-
-        store.commit_transaction(tx_id).await.unwrap();
-
-        let pending_after = store.get_pending_transactions().await.unwrap();
-        assert_eq!(pending_after.len(), 0);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_consistency_validation() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let store = ProductionGraphStore::new(graph, config, resolver);
-
-        let report = store.validate_consistency().await.unwrap();
-        assert!(report.is_consistent);
-        assert_eq!(report.issues_found, 0);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_production_graph_store_event_publishing() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let store = ProductionGraphStore::new(graph, config, resolver);
-        let mut event_receiver = store.subscribe_events();
-
-        let entity = Entity::new(
-            EntityId::new("test_entity".to_string()),
-            "Test Entity".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        // Start a task to upsert entity
-        let store_clone = Arc::new(tokio::sync::Mutex::new(store));
-        let store_for_task = Arc::clone(&store_clone);
-
-        tokio::spawn(async move {
-            let mut store = store_for_task.lock().await;
-            let _ = store.upsert_entity(entity).await;
-        });
-
-        // Wait for event
-        let event =
-            tokio::time::timeout(std::time::Duration::from_millis(100), event_receiver.recv())
-                .await;
-        assert!(event.is_ok());
-    }
-
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn test_incremental_pagerank_creation() {
-        let pagerank = IncrementalPageRank::new(0.85, 1e-6, 100);
-        assert!(pagerank.scores.is_empty());
-    }
-
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn test_batch_processor_creation() {
-        let processor = BatchProcessor::new(100, Duration::from_millis(500), 10);
-        let metrics = processor.get_metrics();
-        assert_eq!(metrics.total_batches_processed, 0);
-    }
-
-    #[cfg(feature = "incremental")]
-    #[tokio::test]
-    async fn test_selective_invalidation() {
-        let invalidation = SelectiveInvalidation::new();
-
-        let region = CacheRegion {
-            region_id: "test_region".to_string(),
-            entity_ids: [EntityId::new("entity1".to_string())].into_iter().collect(),
-            relationship_types: ["KNOWS".to_string()].into_iter().collect(),
-            document_ids: HashSet::new(),
-            last_modified: Utc::now(),
-        };
-
-        invalidation.register_cache_region(region);
-
-        let entity = Entity::new(
-            EntityId::new("entity1".to_string()),
-            "Entity 1".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let ent_id_for_log = entity.id.clone();
-        let change = ChangeRecord {
-            change_id: UpdateId::new(),
-            timestamp: Utc::now(),
-            change_type: ChangeType::EntityUpdated,
-            entity_id: Some(ent_id_for_log),
-            document_id: None,
-            operation: Operation::Update,
-            data: ChangeData::Entity(entity),
-            metadata: HashMap::new(),
-        };
-
-        let strategies = invalidation.invalidate_for_changes(&[change]);
-        assert!(!strategies.is_empty());
-    }
-
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn test_conflict_resolver_merge() {
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-
-        let entity1 = Entity::new(
-            EntityId::new("entity1".to_string()),
-            "Entity 1".to_string(),
-            "Person".to_string(),
-            0.8,
-        );
-
-        let entity2 = Entity::new(
-            EntityId::new("entity1".to_string()),
-            "Entity 1 Updated".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let merged = resolver.merge_entities(&entity1, &entity2).unwrap();
-        assert_eq!(merged.confidence, 0.9); // Should take higher confidence
-        assert_eq!(merged.name, "Entity 1 Updated");
-    }
-
-    #[test]
-    fn test_graph_statistics_creation() {
-        let stats = GraphStatistics {
-            node_count: 100,
-            edge_count: 150,
-            average_degree: 3.0,
-            max_degree: 10,
-            connected_components: 1,
-            clustering_coefficient: 0.3,
-            last_updated: Utc::now(),
-        };
-
-        assert_eq!(stats.node_count, 100);
-        assert_eq!(stats.edge_count, 150);
-    }
-
-    #[test]
-    fn test_consistency_report_creation() {
-        let report = ConsistencyReport {
-            is_consistent: true,
-            orphaned_entities: vec![],
-            broken_relationships: vec![],
-            missing_embeddings: vec![],
-            validation_time: Utc::now(),
-            issues_found: 0,
-        };
-
-        assert!(report.is_consistent);
-        assert_eq!(report.issues_found, 0);
-    }
-
-    #[test]
-    fn test_change_event_creation() {
-        let event = ChangeEvent {
-            event_id: UpdateId::new(),
-            event_type: ChangeEventType::EntityUpserted,
-            entity_id: Some(EntityId::new("entity1".to_string())),
-            timestamp: Utc::now(),
-            metadata: HashMap::new(),
-        };
-
-        assert!(matches!(event.event_type, ChangeEventType::EntityUpserted));
-        assert!(event.entity_id.is_some());
-    }
-
-    // Regression for issue #11: UpdateMonitor::complete_operation must not
-    // self-deadlock when calling update_performance_stats while holding the
-    // operations_log parking_lot Mutex. Runs the call on a dedicated OS
-    // thread and aborts via mpsc::recv_timeout so a deadlock fails the test
-    // (instead of hanging the cargo test process).
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn complete_operation_does_not_self_deadlock() {
-        let monitor = Arc::new(UpdateMonitor::new());
-        let op = monitor.start_operation("regression_11");
-        let m = Arc::clone(&monitor);
-        let (tx, rx) = std::sync::mpsc::channel::<()>();
-        std::thread::spawn(move || {
-            m.complete_operation(&op, true, None, 1, 0);
-            let _ = tx.send(());
-        });
-        rx.recv_timeout(Duration::from_secs(2))
-            .expect("complete_operation self-deadlocked");
-        let stats = monitor.get_performance_stats();
-        assert_eq!(stats.total_operations, 1);
-        assert_eq!(stats.successful_operations, 1);
-    }
-
-    // Regression for issue #13: apply_change_with_conflict_resolution must
-    // return change.change_id (not the monitor's operation_id) so callers
-    // can look the returned id up in change_log without panicking on the
-    // immediate .unwrap() at the upsert_entity call site. Runs the async
-    // call inside an OS thread + dedicated runtime to bound failure time.
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn upsert_entity_returns_id_present_in_change_log() {
-        let graph = KnowledgeGraph::new();
-        let config = IncrementalConfig::default();
-        let resolver = ConflictResolver::new(ConflictStrategy::Merge);
-        let store = Arc::new(tokio::sync::Mutex::new(ProductionGraphStore::new(
-            graph, config, resolver,
-        )));
-        let entity = Entity::new(
-            EntityId::new("regression_13".to_string()),
-            "Regression 13".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-
-        let store_for_thread = Arc::clone(&store);
-        let (tx, rx) = std::sync::mpsc::channel::<Result<UpdateId>>();
-        std::thread::spawn(move || {
-            let rt = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .expect("build runtime");
-            let result = rt.block_on(async {
-                let mut s = store_for_thread.lock().await;
-                s.upsert_entity(entity).await
-            });
-            let _ = tx.send(result);
-        });
-
-        let returned = rx
-            .recv_timeout(Duration::from_secs(2))
-            .expect("upsert_entity hung (likely #11 deadlock or #13 panic)")
-            .expect("upsert_entity returned err");
-
-        let verify_rt = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .expect("build verify runtime");
-        verify_rt.block_on(async {
-            let s = store.lock().await;
-            assert!(
-                s.change_log.get(&returned).is_some(),
-                "returned id is not present in change_log; apply_change_with_conflict_resolution returned the wrong id"
-            );
-        });
-    }
-
-    // Regression for issue #12: BatchProcessor::add_change must release the
-    // pending_batches entry RefMut before calling pending_batches.remove on
-    // the same key, or DashMap shard reentry deadlocks the caller.
-    #[cfg(feature = "incremental")]
-    #[test]
-    fn add_change_does_not_deadlock_on_batch_flush() {
-        let processor = Arc::new(BatchProcessor::new(1, Duration::from_millis(500), 4));
-        let entity = Entity::new(
-            EntityId::new("regression_12".to_string()),
-            "Regression 12".to_string(),
-            "Person".to_string(),
-            0.9,
-        );
-        let change = ChangeRecord {
-            change_id: UpdateId::new(),
-            timestamp: Utc::now(),
-            change_type: ChangeType::EntityAdded,
-            entity_id: Some(entity.id.clone()),
-            document_id: None,
-            operation: Operation::Insert,
-            data: ChangeData::Entity(entity),
-            metadata: HashMap::new(),
-        };
-
-        let p = Arc::clone(&processor);
-        let (tx, rx) = std::sync::mpsc::channel::<Result<String>>();
-        std::thread::spawn(move || {
-            let rt = tokio::runtime::Builder::new_current_thread()
-                .enable_all()
-                .build()
-                .expect("build runtime");
-            let result = rt.block_on(async { p.add_change(change).await });
-            let _ = tx.send(result);
-        });
-
-        let batch_id = rx
-            .recv_timeout(Duration::from_secs(2))
-            .expect("add_change deadlocked on batch flush")
-            .expect("add_change returned err");
-        assert!(!batch_id.is_empty());
-    }
-}
